+++ das-20200116/src/Models/ChatResponseBody.php
@@ -16,6 +16,11 @@ class ChatResponseBody extends Model
     /**
      * @var string
      */
+    public $agentId;
+
+    /**
+     * @var string
+     */
     public $content;
 
     /**
@@ -26,12 +31,32 @@ class ChatResponseBody extends Model
     /**
      * @var string
      */
+    public $kind;
+
+    /**
+     * @var string
+     */
+    public $label;
+
+    /**
+     * @var string
+     */
     public $messageId;
 
     /**
      * @var string
      */
     public $name;
+
+    /**
+     * @var string
+     */
+    public $originatingToolCallId;
+
+    /**
+     * @var string
+     */
+    public $parentAgentId;
 
     /**
      * @var string
@@ -56,12 +81,27 @@ class ChatResponseBody extends Model
     /**
      * @var string
      */
+    public $stepStatus;
+
+    /**
+     * @var string
+     */
     public $taskTrackerId;
 
     /**
      * @var string
      */
     public $threadId;
+
+    /**
+     * @var int
+     */
+    public $timestamp;
+
+    /**
+     * @var string
+     */
+    public $toolCallError;
 
     /**
      * @var string
@@ -84,16 +124,24 @@ class ChatResponseBody extends Model
     public $value;
     protected $_name = [
         'activityType' => 'ActivityType',
+        'agentId' => 'AgentId',
         'content' => 'Content',
         'delta' => 'Delta',
+        'kind' => 'Kind',
+        'label' => 'Label',
         'messageId' => 'MessageId',
         'name' => 'Name',
+        'originatingToolCallId' => 'OriginatingToolCallId',
+        'parentAgentId' => 'ParentAgentId',
         'parentMessageId' => 'ParentMessageId',
         'role' => 'Role',
         'runId' => 'RunId',
         'stepName' => 'StepName',
+        'stepStatus' => 'StepStatus',
         'taskTrackerId' => 'TaskTrackerId',
         'threadId' => 'ThreadId',
+        'timestamp' => 'Timestamp',
+        'toolCallError' => 'ToolCallError',
         'toolCallId' => 'ToolCallId',
         'toolCallName' => 'ToolCallName',
         'type' => 'Type',
@@ -112,6 +160,10 @@ class ChatResponseBody extends Model
             $res['ActivityType'] = $this->activityType;
         }
 
+        if (null !== $this->agentId) {
+            $res['AgentId'] = $this->agentId;
+        }
+
         if (null !== $this->content) {
             $res['Content'] = $this->content;
         }
@@ -120,12 +172,28 @@ class ChatResponseBody extends Model
             $res['Delta'] = $this->delta;
         }
 
+        if (null !== $this->kind) {
+            $res['Kind'] = $this->kind;
+        }
+
+        if (null !== $this->label) {
+            $res['Label'] = $this->label;
+        }
+
         if (null !== $this->messageId) {
             $res['MessageId'] = $this->messageId;
         }
 
         if (null !== $this->name) {
             $res['Name'] = $this->name;
+        }
+
+        if (null !== $this->originatingToolCallId) {
+            $res['OriginatingToolCallId'] = $this->originatingToolCallId;
+        }
+
+        if (null !== $this->parentAgentId) {
+            $res['ParentAgentId'] = $this->parentAgentId;
         }
 
         if (null !== $this->parentMessageId) {
@@ -144,12 +212,24 @@ class ChatResponseBody extends Model
             $res['StepName'] = $this->stepName;
         }
 
+        if (null !== $this->stepStatus) {
+            $res['StepStatus'] = $this->stepStatus;
+        }
+
         if (null !== $this->taskTrackerId) {
             $res['TaskTrackerId'] = $this->taskTrackerId;
         }
 
         if (null !== $this->threadId) {
             $res['ThreadId'] = $this->threadId;
+        }
+
+        if (null !== $this->timestamp) {
+            $res['Timestamp'] = $this->timestamp;
+        }
+
+        if (null !== $this->toolCallError) {
+            $res['ToolCallError'] = $this->toolCallError;
         }
 
         if (null !== $this->toolCallId) {
@@ -183,6 +263,10 @@ class ChatResponseBody extends Model
             $model->activityType = $map['ActivityType'];
         }
 
+        if (isset($map['AgentId'])) {
+            $model->agentId = $map['AgentId'];
+        }
+
         if (isset($map['Content'])) {
             $model->content = $map['Content'];
         }
@@ -191,12 +275,28 @@ class ChatResponseBody extends Model
             $model->delta = $map['Delta'];
         }
 
+        if (isset($map['Kind'])) {
+            $model->kind = $map['Kind'];
+        }
+
+        if (isset($map['Label'])) {
+            $model->label = $map['Label'];
+        }
+
         if (isset($map['MessageId'])) {
             $model->messageId = $map['MessageId'];
         }
 
         if (isset($map['Name'])) {
             $model->name = $map['Name'];
+        }
+
+        if (isset($map['OriginatingToolCallId'])) {
+            $model->originatingToolCallId = $map['OriginatingToolCallId'];
+        }
+
+        if (isset($map['ParentAgentId'])) {
+            $model->parentAgentId = $map['ParentAgentId'];
         }
 
         if (isset($map['ParentMessageId'])) {
@@ -215,12 +315,24 @@ class ChatResponseBody extends Model
             $model->stepName = $map['StepName'];
         }
 
+        if (isset($map['StepStatus'])) {
+            $model->stepStatus = $map['StepStatus'];
+        }
+
         if (isset($map['TaskTrackerId'])) {
             $model->taskTrackerId = $map['TaskTrackerId'];
         }
 
         if (isset($map['ThreadId'])) {
             $model->threadId = $map['ThreadId'];
+        }
+
+        if (isset($map['Timestamp'])) {
+            $model->timestamp = $map['Timestamp'];
+        }
+
+        if (isset($map['ToolCallError'])) {
+            $model->toolCallError = $map['ToolCallError'];
         }
 
         if (isset($map['ToolCallId'])) {